+++ src/server.rs
@@ -94,7 +94,7 @@ pub enum Package {
     Version {
         addr_from: String,
         version: u32,
-        best_height: u32,
+        best_height: u64,
     },
 }
 
@@ -148,7 +148,7 @@ pub fn send_tx(addr: &str, tx: &Transaction) {
     );
 }
 
-fn send_version(addr: &str, height: u32) {
+fn send_version(addr: &str, height: u64) {
     let socket_addr = addr.parse().unwrap();
     let node_addr = GLOBAL_CONFIG.get_node_addr().parse().unwrap();
     send_data(
